CRUD Javafx Application using MySQL
This application uses JavaFX to connect to a MySQL database.  The application pulls data from the database, creates a table to hold the data, and enables the user to manipulate the data through the UI.  This project was created to input and manipulate data in a book log efficiently and easily rather than having to keep up with a spreadsheet.